Locking down delete permissions

+++ backend/src/resolvers/Mutation.js
@@ -46,10 +46,14 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    // const item = await ctx.db.query.item({ where }, info);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
 
     // 2. Check if they own that item, or have the permissions
-    // TODO
+    const ownsItems = item.user.id === ctx.request.userId;
+    const hasDeletePermissions = ctx.request.user.permissions.some(p => ['ADMIN', 'ITEMDELETE'.includes(p)]);
+    if (!ownsItems || !hasDeletePermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
 
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
